backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
langchain-community
langchain-huggingface
langchain-chroma
chromadb
openai
spacy
pandas
pypdf
pdfplumber
sentence-transformers
python-dotenv
aiofiles
xlrd
openpyxl
Jinja2
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
transformers>=4.30.0
torch>=2.0.0
PyMuPDF
numpy
